requirements.txt
black>=23.7.0
graphviz>=0.20.1
inflection>=0.5.1
mapomatic>=0.10.0
matplotlib~=3.8.2
mqt.bench>=1.0.3
mqt.predictor>=1.3.0
numpy>=1.23.5
pandas>=2.2.0
pymoo>=0.6.0.1
pytest>=7.4.1
pytest-profiling>=1.7.0
qiskit>=0.44.1
qiskit-ibm-provider>=0.7.1
qiskit-ibm-runtime>=0.12.0
retry>=0.9.2
scipy>=1.11.2
seaborn>=0.12.2
Supermarq>=0.4.17
SQLAlchemy>=2.0.20
joblib>=1.3.2
scikit-learn>=1.3.0
jsonpickle>=3.0.2
qiskit_qasm3_import>=0.4.2
